refactor(moe): move AMD latent registrations
