+++ tokenspeed-kernel/test/ops/moe/test_latent_input.py
@@ -6,6 +6,8 @@ import pytest
 import torch
 from tokenspeed_kernel.ops.moe import latent_moe_input_projections
 from tokenspeed_kernel.ops.moe.latent_input import packed_projection_weight_view
+from tokenspeed_kernel.platform import current_platform
+from tokenspeed_kernel.registry import KernelRegistry
 
 if not torch.cuda.is_available():
     pytest.skip("requires a GPU", allow_module_level=True)
@@ -16,6 +18,31 @@ ROUTED_N = 512
 SHARED_N = 128
 GATE_CLAMP = 2.5
 UP_CLAMP = 1.0
+
+
+def test_amd_latent_moe_registrations_are_platform_scoped() -> None:
+    registry = KernelRegistry.get()
+    expected_modules = {
+        "gluon_latent_expert_shared_gfx950": (
+            "tokenspeed_kernel_amd.ops.gfx950.moe.latent_decode"
+        ),
+        "gluon_latent_input_decode_gfx950": (
+            "tokenspeed_kernel_amd.ops.gfx950.moe.latent_input"
+        ),
+        "gluon_latent_input_small_batch_gfx950": (
+            "tokenspeed_kernel_amd.ops.gfx950.moe.latent_input"
+        ),
+    }
+    implementations = {name: registry.get_impl(name) for name in expected_modules}
+    if current_platform().is_amd:
+        assert {
+            name: implementation.__module__
+            for name, implementation in implementations.items()
+        } == expected_modules
+    else:
+        assert all(
+            implementation is None for implementation in implementations.values()
+        )
 
 
 def _concatenated_weights(
